refactor: move default location into consts file
