+++ src/utils/consts.ts
@@ -1,5 +1,13 @@
 export const ADDRESS_FINDER_API_KEY = "ARFHPVK67QXM49BEWDL3";
+
 export const REVERSE_GEOCODE_URL =
   "https://api.addressfinder.io/api/nz/address/reverse_geocode/?";
+
 export const ADDRESS_METADATA_URL =
   "https://api.addressfinder.io/api/nz/address/metadata/?";
+
+export const DEFAULT_LOCATION = {
+  lat: -36.853610199274385,
+  lng: 174.76054541484535,
+  placeName: "Auckland CBD",
+};
