--- src/utils/location.ts
@@ -1,5 +1,0 @@
-export const DEFAULT_LOCATION = {
-  lat: -36.853610199274385,
-  lng: 174.76054541484535,
-  placeName: "Auckland CBD",
-};
